+++ src/main.js
@@ -1,8 +1,4 @@
-import { FoodForm } from "./component/FoodForm.js";
-import { FoodItem } from "./component/FoodItem.js";
 import { FoodList } from "./component/FoodList.js";
-import { HeaderComponent } from "./component/HeaderComponent.js";
-import { Modal } from "./component/Modal.js";
 import { foodItems } from "./mock/mockItem.js";
 import { FoodListPage } from "./pages/FoodListPage.js";
